Upgrade ubuntu versions of docker images to 18.04
Upgrading the mysql version in the multi process container was a massive
pain. With version 5.7 mysql per-default requires the root user to be
logged in from a root shell. Since we start the container as non-root we
have to set the password of the mysql root user in a initialization
script and disable the `auth_socket` login method.